Set ARM thumb enabled if bit zero of PC is set
When the program counter is set the current code masks out
the bottom bit. The general intention when setting the low
bit on a code address on ARM is to enable thumb mode, so
emulate this behaviour in the gdbstub. Ideally check if
a similar behaviour occurs on real hardware.